Fixing record allocation for large grids.

+++ DBlib/src/DBImpASCIINet.C
@@ -114,7 +114,7 @@ int DBImportASCIINet(DBObjData *netData, const char *fileName) {
     valueSizeFLD->Int(layerRec, sizeof(DBInt));
     rowNumFLD->Int(layerRec, rowNum);
     colNumFLD->Int(layerRec, colNum);
-    if ((dataRec = new DBObjRecord("NetLookupGridRecord", ((size_t) rowNum) * colNum * sizeof(DBInt), sizeof(DBInt))) ==
+    if ((dataRec = new DBObjRecord("NetLookupGridRecord", (size_t) rowNum * (size_t) colNum * sizeof(DBInt), sizeof(DBInt))) ==
         (DBObjRecord *) NULL)
         return (DBFault);
     layerFLD->Record(layerRec, dataRec);
